Task 13 done

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -16,9 +16,11 @@ def main():
     engine = create_engine(conn)
     Session = sessionmaker(bind=engine)
     session = Session()
-    for instance in session.query(State).filter(State.name.like('%a%')):
-        session.delete(instance)
-        session.commit()
+    rows = session.query(State).filter(State.name.like('%a%')).all()
+    if rows:
+       for instance in rows:
+            session.delete(instance)
+            session.commit()
     session.close()
